@@ -1,26 +1,29 @@
 plugins {
-    id 'org.springframework.boot' version '2.1.6.RELEASE'
+    id 'org.springframework.boot' version '2.2.2.RELEASE'
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
     id 'java'
 }
 
-apply plugin: 'io.spring.dependency-management'
-
-group = 'lk.imms'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
 
 configurations {
     developmentOnly
     runtimeClasspath {
         extendsFrom developmentOnly
     }
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
 }
 
 repositories {
     mavenCentral()
+}
+
+springBoot {
+    buildInfo {
+        properties {
+            group = 'com.example'
+            version = '0.0.1-SNAPSHOT'
+            sourceCompatibility = '11'
+        }
+    }
 }
 
 dependencies {
