@@ -6,17 +6,17 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public enum PetitionType {
+    PRESIDENT("By President Office"),
+    RRIMEMINISTER("By Prime Minister Office"),
+    NDDCA("By NDDCA"),
     TELEPHONE("By Telephoning"),
     LETTERS("By Letters"),
     EMAIL("By Email"),
     SMS("By SMS"),
-    FACEBOOK("By Facebook"),
-    FORMWEB("By Web Site"),
-    RESIDENT("By President Office"),
-    RRIMEMINISTER("By Prime Minister Office"),
-    NDDCA("By NDDCA"),
     OTHERAUTHORITIES("By Other Autho"),
-    OTHER("By Other");
+    OTHER("By Direct Handover"),
+    FACEBOOK("By Facebook"),
+    FORMWEB("By Web Site");
 
     private final String petitionType;
 }
