@@ -1,44 +1,25 @@
 package lk.imms.management_system.asset.petition.entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lk.imms.management_system.asset.petition.entity.Enum.PetitionStateType;
-import lk.imms.management_system.general.security.entity.User;
+import lk.imms.management_system.util.audit.AuditEntity;
 import lombok.*;
-import org.springframework.format.annotation.DateTimeFormat;
 
-import javax.persistence.*;
-import java.time.LocalDate;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.ManyToOne;
 
 @Entity
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@EqualsAndHashCode
-@JsonIgnoreProperties( value = {"createdAt", "updatedAt"}, allowGetters = true )
-public class PetitionState {
-    @Id
-    @GeneratedValue( strategy = GenerationType.IDENTITY )
-    @Column( unique = true )
-    private Long id;
+public class PetitionState extends AuditEntity {
 
     @Enumerated( EnumType.STRING )
     private PetitionStateType petitionStateType;
 
-    private String remark;
-
     @ManyToOne
     private Petition petition;
 
-    @ManyToOne
-    private User createdUser;
-
-    @ManyToOne
-    private User updatedUser;
-
-    @DateTimeFormat( pattern = "yyyy-MM-dd" )
-    private LocalDate createdAt;
-
-    @DateTimeFormat( pattern = "yyyy-MM-dd" )
-    private LocalDate updatedAt;
 }
