@@ -1,6 +1,7 @@
 package lk.imms.management_system.configuration;
 
-import lk.imms.management_system.general.security.service.UserDetailsServiceImpl;
+
+import lk.imms.management_system.asset.userManagement.service.UserDetailsServiceImpl;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
@@ -8,15 +9,15 @@ import org.springframework.security.config.annotation.authentication.builders.Au
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 @Configuration
 @EnableWebSecurity
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
-    /*@Autowired
-    public UserDetailsServiceImpl userDetailsService;*/
+    private final String[] ALL_PERMIT_URL = {"/static/favicon.ico", "/img/**", "/css/**", "/js/**", "/webjars/**",
+            "/actuator/**", "/login", "/select/**", "/", "/index"};
 
     @Bean
     public UserDetailsServiceImpl userDetailsService() {
@@ -37,64 +38,58 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     }
 
     @Override
-    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+    protected void configure(AuthenticationManagerBuilder auth) {
         auth.authenticationProvider(authenticationProvider());
     }
 
 
-/*    @Override
-    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.userDetailsService(userDetailsService).passwordEncoder(passwordEncoder);
-    }*/
-
     @Override
     protected void configure(HttpSecurity http) throws Exception {
-/*        http.csrf().disable();
-        http.authorizeRequests().antMatchers("/").permitAll();*/
+     /*       http.csrf().disable();
+            http.authorizeRequests().antMatchers("/").permitAll();
+    */
+        // For developing easy to give permission all lin
 
-        //for developing easy to give permission all link
-
-        http.
-                authorizeRequests()
-                //Always users can access without login
-                .antMatchers(
-                        "/index",
-                        "/favicon.ico",
-                        "/img/**",
-                        "/css/**",
-                        "/js/**",
-                        "/fonts/**",
-                        "/fontawesome/**").permitAll()
-                .antMatchers("/login", "/select/**").permitAll()
-
-                //Need to login for access those are
-                .antMatchers("/employee/**").hasRole("MANAGER")
-                .antMatchers("/employee/**").hasRole("MANAGER")
-                .antMatchers("/user/**").hasRole("MANAGER")
-                .antMatchers("/invoiceProcess/add").hasRole("CASHIER")
-                .anyRequest()
-                .authenticated()
-                .and()
+        http
+                .authorizeRequests(
+                        authorizeRequests ->
+                                authorizeRequests
+                                        //Anytime users can access without login
+                                        //to see actuator details
+                                        .antMatchers(ALL_PERMIT_URL).permitAll()
+                                        //this is used the normal admin to give access every url mapping
+                                        .antMatchers("/").hasRole("/ADMIN")
+                                        //Need to login for access those are
+                                        .antMatchers("/employee/**").hasRole("ADMIN")
+                                        .antMatchers("/employee1/**").hasRole("MANAGER")
+                                        .antMatchers("/user/**").hasRole("ADMIN")
+                                        .antMatchers("/petition/**").hasRole("ADMIN")
+                                        .antMatchers("/minutePetition/**").hasRole("MANAGER")
+                                        .antMatchers("/invoiceProcess/add").hasRole("CASHIER")
+                                        .anyRequest()
+                                        .authenticated())
                 // Login form
-                .formLogin()
-                .loginPage("/login")
-                .defaultSuccessUrl("/mainWindow")
-                //Username and password for validation
-                .usernameParameter("userame")
-                .passwordParameter("password")
-                .and()
+                .formLogin(
+                        formLogin ->
+                                formLogin
+                                        .loginPage("/login")
+                                        //Username and password for validation
+                                        .usernameParameter("username")
+                                        .passwordParameter("password")
+                                        .defaultSuccessUrl("/mainWindow"))
                 //Logout controlling
-                .logout()
-                .invalidateHttpSession(true)
-                .clearAuthentication(true)
-                .logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
-                .logoutSuccessUrl("/index")
-                .and()
-                .exceptionHandling()
+                .logout(
+                        logout ->
+                                logout
+                                        .logoutUrl("/logout")
+                                        .deleteCookies("JSESSIONID")
+                                        .invalidateHttpSession(false)
+                                        .clearAuthentication(true)
+                                        .logoutSuccessUrl("/login"))
                 //Cross site disable
-                .and()
-                .csrf()
-                .disable();
-
+                .csrf(AbstractHttpConfigurer::disable)
+                .exceptionHandling();
     }
+
 }
+
